Encapsulate translation of pattern into DFA
This is in preparation for introducing alternative matching
strategies that are no so expensive to compile.
