+++ core/trino-main/src/main/java/io/trino/likematcher/DenseDfaMatcher.java
@@ -13,6 +13,11 @@
  */
 package io.trino.likematcher;
 
+import java.util.List;
+
+import static com.google.common.base.Preconditions.checkArgument;
+import static java.nio.charset.StandardCharsets.UTF_8;
+
 class DenseDfaMatcher
 {
     // The DFA is encoded as a sequence of transitions for each possible byte value for each state.
@@ -35,8 +40,10 @@ class DenseDfaMatcher
     /**
      * @param exact whether to match to the end of the input
      */
-    public static DenseDfaMatcher newInstance(DFA dfa, boolean exact)
+    public static DenseDfaMatcher newInstance(List<Pattern> pattern, boolean exact)
     {
+        DFA dfa = makeNfa(pattern).toDfa();
+
         int[] transitions = new int[dfa.states().size() * 256];
         boolean[] accept = new boolean[dfa.states().size()];
 
@@ -110,5 +117,80 @@ class DenseDfaMatcher
         }
 
         return accept[state >>> 8];
+    }
+
+    private static NFA makeNfa(List<Pattern> pattern)
+    {
+        checkArgument(!pattern.isEmpty(), "pattern is empty");
+
+        NFA.Builder builder = new NFA.Builder();
+
+        NFA.State state = builder.addStartState();
+
+        for (Pattern item : pattern) {
+            if (item instanceof Pattern.Literal literal) {
+                for (byte current : literal.value().getBytes(UTF_8)) {
+                    state = matchByte(builder, state, current);
+                }
+            }
+            else if (item instanceof Pattern.Any any) {
+                for (int i = 0; i < any.min(); i++) {
+                    NFA.State next = builder.addState();
+                    matchSingleUtf8(builder, state, next);
+                    state = next;
+                }
+
+                if (any.unbounded()) {
+                    matchSingleUtf8(builder, state, state);
+                }
+            }
+            else {
+                throw new UnsupportedOperationException("Not supported: " + item.getClass().getName());
+            }
+        }
+
+        builder.setAccept(state);
+
+        return builder.build();
+    }
+
+    private static NFA.State matchByte(NFA.Builder builder, NFA.State state, byte value)
+    {
+        NFA.State next = builder.addState();
+        builder.addTransition(state, new NFA.Value(value), next);
+        return next;
+    }
+
+    private static void matchSingleUtf8(NFA.Builder builder, NFA.State from, NFA.State to)
+    {
+        /*
+            Implements a state machine to recognize UTF-8 characters.
+
+                  11110xxx       10xxxxxx       10xxxxxx       10xxxxxx
+              O ───────────► O ───────────► O ───────────► O ───────────► O
+              │                             ▲              ▲              ▲
+              ├─────────────────────────────┘              │              │
+              │          1110xxxx                          │              │
+              │                                            │              │
+              ├────────────────────────────────────────────┘              │
+              │                   110xxxxx                                │
+              │                                                           │
+              └───────────────────────────────────────────────────────────┘
+                                        0xxxxxxx
+        */
+
+        builder.addTransition(from, new NFA.Prefix(0, 1), to);
+
+        NFA.State state1 = builder.addState();
+        NFA.State state2 = builder.addState();
+        NFA.State state3 = builder.addState();
+
+        builder.addTransition(from, new NFA.Prefix(0b11110, 5), state1);
+        builder.addTransition(from, new NFA.Prefix(0b1110, 4), state2);
+        builder.addTransition(from, new NFA.Prefix(0b110, 3), state3);
+
+        builder.addTransition(state1, new NFA.Prefix(0b10, 2), state2);
+        builder.addTransition(state2, new NFA.Prefix(0b10, 2), state3);
+        builder.addTransition(state3, new NFA.Prefix(0b10, 2), to);
     }
 }
